+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * #23 Fix Select with Addons overflowing
 * Add ``loading`` property to Select
+* #30 Fix Table backend-sorting resetting sorted column
 
 ## 0.2.0