+++ CHANGELOG.md
@@ -2,10 +2,12 @@
 
 ## 0.2.1
 
-* #23 Fix Select with Addons overflowing
+* #23 Fix Select with addons overflowing
 * Add ``loading`` property to Select
 * #30 Fix Table ``backend-sorting`` resetting sorted column
 * Fix Checkbox disabled style
+* Fix Input ``loading`` overlapping other icons
+* Fix Input not firing HTML5 validation when v-model is updated
 
 ## 0.2.0